Backend/requirements.txt
Flask-SQLAlchemy>=2.1
flask-marshmallow>=0.7.0
marshmallow-sqlalchemy>=0.10.0
flask-restplus>=0.9.2
Flask-Migrate>=2.0.0
Flask-Script>=2.0.5

